+++ changelog.md
@@ -1,3 +1,6 @@
+# SDK v.2.5.10
+- Add login method using nonce and JWT identity_token
+
 # SDK v.2.5.9
 - Add file_attachment message type rendering
 - `QiscusSDK.core.searchMessages` API